remove blog feature (in favor of local gatsby plugin in project)
