--- example/content/en/authors/example-author.md
@@ -1,3 +1,0 @@
----
-name: Example author
----
