--- example/content/en/categories/polkadot.md
@@ -1,7 +1,0 @@
----
-title: Polkadot
-slug: polkadot
-index: 1
----
-
-The Polkadot category in the blog.
