--- example/content/en/categories/web3.md
@@ -1,7 +1,0 @@
----
-title: web3
-slug: web3
-index: 0
----
-
-The web3 category.
