--- example/content/en/posts/1.md
@@ -1,13 +1,0 @@
----
-title: Introduction (First)
-slug: ''
-created_at: 2021-07-27T15:09:14.509Z
-featured_image: /media/favicon/favicon.png
-tags:
-  - test
----
-# Introduction article for the blog.
-
-<w3f-link-cta href="https://example.com">Example site CTA</w3f-link-cta>
-
-A sample article
