--- example/content/en/posts/2.md
@@ -1,13 +1,0 @@
----
-title: Second
-slug: ''
-created_at: 2021-07-28T15:09:14.509Z
-featured_image: /media/favicon/favicon.png
-tags:
-  - w3f
-  - test
-is_featured: true
----
-# A second introduction article for the blog.
-
-A second sample article
