--- example/content/en/posts/3.md
@@ -1,13 +1,0 @@
----
-title: Third
-slug: 'third-test-post'
-created_at: 2021-07-29T15:09:14.509Z
-featured_image: /media/favicon/favicon.png
-tags:
-  - w3f
-categories:
-  - web3
----
-# Third article for the blog.
-
-A sample article, as 3rd element.
