--- example/content/en/tags/test.md
@@ -1,4 +1,0 @@
----
-slug: test
-title: test
----
