--- example/content/en/tags/w3f.md
@@ -1,4 +1,0 @@
----
-slug: w3f
-title: w3f
----
