--- example/src/@w3f/gatsby-theme-w3f/config/menus.js
@@ -1,16 +1,0 @@
-/* the main menu, ids of items must match
-   the submenu's key of this js object */
-const main = [
-  {
-    url: '/',
-    id: 'home',
-  },
-  {
-    url: '/blog',
-    id: 'blog',
-  }
-];
-
-module.exports = {
- main
-};
