--- example/src/@w3f/gatsby-theme-w3f/hooks/use-site-menus.js
@@ -1,23 +1,0 @@
-import { graphql, useStaticQuery } from 'gatsby';
-
-export const useSiteMenus = () => {
-  const {
-    site: {
-      siteMetadata: { menus },
-    },
-  } = useStaticQuery(graphql`
-    query {
-      site {
-        siteMetadata {
-          menus {
-            main {
-              id
-              url
-            }
-          }
-        }
-      }
-    }
-  `);
-  return { menus };
-};
