+++ gatsby-theme-w3f/gatsby-node.js
@@ -1,6 +1,31 @@
-const { addLangFieldToMarkdown, addSlugFieldToMarkdown } = require('./gatsby-node/on-create-node.js');
+const { createFilePath } = require('gatsby-source-filesystem');
 
-const { createBlogPages, createBlogTagsPages } = require('./gatsby-node/create-pages.js');
+const addLangFieldToMarkdown = ({ actions, getNode, node }) => {
+  /* Add a `langKey` field on each node;
+     based on which "root /content" folder, the nodes are in */
+  const { createNodeField } = actions;
+  const fileNode = getNode(node.parent);
+  const langKey = fileNode.relativePath.split('/')[0];
+  createNodeField({
+    node,
+    name: `langKey`,
+    value: langKey,
+  });
+};
+
+const addSlugFieldToMarkdown = ({ node, getNode, actions }) => {
+  const { createNodeField } = actions;
+  const filepaths = createFilePath({ node, getNode, basePath: `pages` })
+    .split('/')
+    .filter(pathElement => pathElement);
+  const slugPath = filepaths[filepaths.length - 1];
+  const slug = node.frontmatter && node.frontmatter.slug ? node.frontmatter.slug : slugPath;
+  createNodeField({
+    node,
+    name: `slug`,
+    value: slug,
+  });
+};
 
 exports.onCreateNode = props => {
   const { node } = props;
@@ -8,9 +33,4 @@ exports.onCreateNode = props => {
     addLangFieldToMarkdown(props);
     addSlugFieldToMarkdown(props);
   }
-};
-
-exports.createPages = async props => {
-  await Promise.all([createBlogPages(props)]);
-  await Promise.all([createBlogTagsPages(props)]);
 };
