--- gatsby-theme-w3f/src/components/tags/Card.js
@@ -1,18 +1,0 @@
-import React from 'react';
-
-import { Link } from '../default/Link';
-
-const TagCard = ({ model }) => {
-  const { frontmatter, fields } = model;
-  const { name } = frontmatter;
-  const { slug } = fields;
-  return (
-    <article>
-      <Link to={`/tags/${slug}`}>
-        <h2>{name}</h2>
-      </Link>
-    </article>
-  );
-};
-
-export default TagCard;
