--- gatsby-theme-w3f/src/components/tags/Single.js
@@ -1,15 +1,0 @@
-import React from 'react';
-
-const TagSingle = ({ model }) => {
-  const { frontmatter } = model;
-  const { title } = frontmatter;
-  return (
-    <article>
-      <header>
-        <h1>{title}</h1>
-      </header>
-    </article>
-  );
-};
-
-export default TagSingle;
